Name Change Form
Added the List Name Change function to a form so the user can press
enter to complete as well as click the button.

List Name input box will now populate from the current List Name when
the edit button is clicked.

Changed opacity on the remove button to make it look gray.